backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.4
python-dotenv==1.0.1